Flask-Framework/requirements.txt
flask-cors==6.0.0
requests==2.32.3
pillow==11.2.1
numpy==2.1.3
tensorflow==2.19.0
ncps==1.0.1